restApi/requirements.txt
aiohttp~=3.6.2
aiohttp-apispec==2.2.3
aiomisc~=9.6.11
alembic~=1.3.3
asyncpgsa==0.27.1
ConfigArgParse~=1.0
psycopg2-binary==2.8.4
pytz==2019.3
setproctitle==1.1.10
SQLAlchemy==1.3.14
python-jose[cryptography]==3.3.0  
passlib[bcrypt]==1.7.4  
PyJWT==2.8.0  
bcrypt==4.0.1 
validators==0.22.0
# Для проверок
requests==2.31.0  # Для HTTP и PING
dnspython==2.4.2  # Для DNS  